bot/requirements.txt
tweepy==3.10.0
sqlalchemy
python-dotenv
joblib
requests
click